refact: Update for rust 1.0.0 alpha

+++ src/mob/mod.rs
@@ -11,28 +11,28 @@ pub fn rnd() -> Mob {
     let rx  = dice::rand(1, 80); //FIXME
     let ry  = dice::rand(1, 50);
 
-    match FromPrimitive::from_int(rnd as int) {
+    match FromPrimitive::from_int(rnd as isize) {
         Some(kind) => match kind {
             Kind::Canine => Mob {
                 name: "Fido".to_string(),
                 pos: Cell::new(Pos(rx, ry)),
                 kind: kind,
                 hp: Cell::new(7), str: 2,
-                ap: 1, int: 7, con: 7, dex: 7, display_char: kind.to_char(), behavior: behavior::Kind::Animalic
+                ap: 1, isize: 7, con: 7, dex: 7, display_char: kind.to_char(), behavior: behavior::Kind::Animalic
             },
             Kind::Hobgoblin => Mob {
                 name: "Gardhur".to_string(),
                 pos: Cell::new(Pos(rx, ry)),
                 kind: kind,
                 hp: Cell::new(12), str: 5,
-                ap: 1, int: 7, con: 7, dex: 7, display_char: kind.to_char(), behavior: behavior::Kind::Animalic
+                ap: 1, isize: 7, con: 7, dex: 7, display_char: kind.to_char(), behavior: behavior::Kind::Animalic
             },
             Kind::Orc => Mob {
                 name: "Gardhur".to_string(),
                 pos: Cell::new(Pos(rx, ry)),
                 kind: kind,
                 hp: Cell::new(15), str: 3,
-                ap: 1, int: 7, con: 7, dex: 7, display_char: kind.to_char(), behavior: behavior::Kind::Animalic
+                ap: 1, isize: 7, con: 7, dex: 7, display_char: kind.to_char(), behavior: behavior::Kind::Animalic
             },
             _ => panic!("Can't spawn a hero")
         },
@@ -48,7 +48,7 @@ pub struct Mob {
     pub name: String,
     pub kind: Kind,
     pub str: u32,
-    int: u32,
+    isize: u32,
     con: u32,
     dex: u32,
     ap: u32,
@@ -65,7 +65,7 @@ impl<'a> Mob {
             ap: 1, 
             hp: Cell::new(20),
             str: 7,
-            int: 7,
+            isize: 7,
             con: 7,
             dex: 7,
             display_char: kind.to_char(),
@@ -99,9 +99,9 @@ impl<'a> Mob {
 
 #[derive(Copy, Clone, Show, PartialEq, FromPrimitive)]
 pub enum Kind {
-    Hero,      //= '@' as int,
-    Canine,    //= 'C' as int,
-    Hobgoblin,  //= 'h' as int
+    Hero,      //= '@' as isize,
+    Canine,    //= 'C' as isize,
+    Hobgoblin,  //= 'h' as isize
     Orc
 }
